requirements.txt
Flask
flask-cors
python-dotenv
openai
flask_sqlalchemy
bcrypt
